♻️ (#80) mock data 분리
