+++ src/features/project/api/projectApi.ts
@@ -1,28 +1,7 @@
 import type { Id } from '@/shared/types/commonTypes';
 import { generateId } from '@/shared/utils/idUtils';
 import type { Project } from '@/features/project/types/projectTypes';
-
-// 임시 mock project 데이터
-let mockProjects: Project[] = [
-  {
-    projectId: 123,
-    name: '팀플 A',
-    defaultReviewerCount: 2,
-    role: 'admin',
-  },
-  {
-    projectId: 124,
-    name: '팀플 B',
-    defaultReviewerCount: 2,
-    role: 'member',
-  },
-  {
-    projectId: 125,
-    name: '팀플 C',
-    defaultReviewerCount: 2,
-    role: 'member',
-  },
-];
+import { mockProjects } from '@/shared/data/mockProjects';
 
 // 추후 실제 API 호출 예정
 export const projectApi = {
@@ -60,7 +39,10 @@ export const projectApi = {
 
   // 프로젝트 삭제
   deleteProject: async (projectId: Id): Promise<{ success: boolean }> => {
-    mockProjects = mockProjects.filter((project) => project.projectId !== projectId);
+    const index = mockProjects.findIndex((project) => project.projectId === projectId);
+    if (index !== -1) {
+      mockProjects.splice(index, 1);
+    }
     return { success: true };
   },
 };
